requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.0.6
django-debug-toolbar==4.4.5
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
sqlparse==0.5.0
tzdata==2024.1
